Create FrakenCrow
This code lights two sets of LEDs with a force sensor. One set of LED's are on a timer and turn on when the force sensor is sensed, the other set blink on and off once the sensor has been touched. The servo motor turns on when the light sensor has been detected.